# openapi.yaml
openapi: 3.0.0
info:
  title: REST API
  description: The REST API is used for all API interactions with DataCite services.
  contact:
    name: DataCite Support
    url: 'https://support.datacite.org'
    email: support@datacite.org
  license:
    name: MIT
    url: 'https://raw.githubusercontent.com/datacite/lupo/master/LICENSE'
  version: 2.3.0
servers:
  - url: 'https://api.test.datacite.org'
    description: Staging server
  - url: 'https://api.datacite.org'
    description: Production server
...
paths:
...
  /providers:
    get:
      tags:
        - providers
      summary: A JSON array of providers.
      security: []
      parameters:
        - in: query
          name: query
          schema:
            type: string
        - in: query
          name: year
          schema:
            type: number
        - in: query
          name: region
          schema:
            type: string
            enum:
              - amer
              - apac
              - emea
        - in: query
          name: organization-type
          schema:
            type: string
            enum:
              - academicInstitution
              - academicLibrary
              - governmentAgency
              - nationalInstitution
              - nationalLibrary
              - publisher
              - professionalSociety
              - researchInstitution
              - serviceProvider
              - vendor
        - in: query
          name: focus-area
          schema:
            type: string
            enum:
              - amer
              - apac
              - emea
        - in: query
          name: 'page[number]'
          schema:
            type: number
        - in: query
          name: 'page[size]'
          schema:
            type: number
        - in: query
          name: sort
          schema:
            type: string
            enum:
              - relevance
              - name
              - '-name'
              - created
              - '-created'
      responses:
        '200':
          description: A JSON array of prefixes.
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/Provider'
    post:
      tags:
        - providers
      summary: Add a new provider.
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              $ref: '#/components/schemas/Provider'
      responses:
        '201':
          description: Created
...
components:
  schemas:
...
    Provider:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            symbol:
              type: string
      required:
        - id
        - attributes - name - symbol
      example:
        id: bl
        attributes:
          name: British Library
          symbol: BL
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BasicAuth: []
  - bearerAuth: []